Move various parts of provider to different files
